Enum quick_junit::TestCaseStatus

pub enum TestCaseStatus {
    Success {
        flaky_runs: Vec<TestRerun>,
    },
    NonSuccess {
        kind: NonSuccessKind,
        message: Option<String>,
        ty: Option<String>,
        description: Option<String>,
        reruns: Vec<TestRerun>,
    },
    Skipped {
        message: Option<String>,
        ty: Option<String>,
        description: Option<String>,
    },
}

Represents the success or failure of a test case.

Variants

Success

Fields

flaky_runs: Vec<TestRerun>

Prior runs of the test. These are represented as flakyFailure or flakyError in the JUnit XML.

This test case passed.

NonSuccess

Fields

kind: NonSuccessKind

Whether this test case failed in an expected way (failure) or an unexpected way (error).

message: Option<String>

The failure message.

ty: Option<String>

The “type” of failure that occurred.

description: Option<String>

The description of the failure.

This is serialized and deserialized from the text node of the element.

reruns: Vec<TestRerun>

Test reruns. These are represented as rerunFailure or rerunError in the JUnit XML.

This test case did not pass.

Skipped

Fields

message: Option<String>

The skip message.

ty: Option<String>

The “type” of skip that occurred.

description: Option<String>

The description of the skip.

This is serialized and deserialized from the text node of the element.

This test case was not run.

Implementations

impl TestCaseStatus

pub fn success() -> Self

Creates a new TestCaseStatus that represents a successful test.

pub fn non_success(kind: NonSuccessKind) -> Self

Creates a new TestCaseStatus that represents an unsuccessful test.

pub fn skipped() -> Self

Creates a new TestCaseStatus that represents a skipped test.

pub fn set_message(&mut self, message: impl Into<String>) -> &mut Self

Sets the message. No-op if this is a success case.

pub fn set_type(&mut self, ty: impl Into<String>) -> &mut Self

Sets the type. No-op if this is a success case.

pub fn set_description(&mut self, description: impl Into<String>) -> &mut Self

Sets the description (text node). No-op if this is a success case.

pub fn add_rerun(&mut self, rerun: TestRerun) -> &mut Self

Adds a rerun or flaky run. No-op if this test was skipped.

pub fn add_reruns(
&mut self,
reruns: impl IntoIterator<Item = TestRerun>
) -> &mut Self

Adds reruns or flaky runs. No-op if this test was skipped.