Enum quick_junit::TestcaseStatus

pub enum TestcaseStatus {
    Success,
    Failure {
        message: Option<String>,
        ty: Option<String>,
        description: Option<String>,
    },
    Error {
        message: Option<String>,
        ty: Option<String>,
        description: Option<String>,
    },
    Skipped {
        message: Option<String>,
        ty: Option<String>,
        description: Option<String>,
    },
}

Represents the success or failure of a testcase.

Variants

Success

This testcase passed.

Failure

This testcase failed in an expected way.

Fields of Failure

message: Option<String>

The failure message.

ty: Option<String>

The "type" of failure that occurred.

description: Option<String>

The description of the failure.

This is serialized and deserialized from the text node of the element.

Error

This testcase errored in an unexpected way.

Fields of Error

message: Option<String>

The error message.

ty: Option<String>

The "type" of error that occurred.

description: Option<String>

The description of the failure.

This is serialized and deserialized from the text node of the element.

Skipped

This testcase was not run.

Fields of Skipped

message: Option<String>

The skip message.

ty: Option<String>

The "type" of skip that occurred.

description: Option<String>

The description of the skip.

This is serialized and deserialized from the text node of the element.

Implementations

impl TestcaseStatus

pub fn success() -> Self

Creates a new TestcaseStatus that represents a successful test.

pub fn failure() -> Self

Creates a new TestcaseStatus that represents a failure.

pub fn error() -> Self

Creates a new TestcaseStatus that represents an error.

pub fn skipped() -> Self

Creates a new TestcaseStatus that represents a skipped test.

pub fn set_message(&mut self, message: impl Into<String>) -> &mut Self

Sets the message. No-op if this is a success case.

pub fn set_type(&mut self, ty: impl Into<String>) -> &mut Self

Sets the type. No-op if this is a success case.

pub fn set_description(&mut self, description: impl Into<String>) -> &mut Self

Sets the description (text node). No-op if this is a success case.