Struct quick_junit::Report

pub struct Report {
    pub name: String,
    pub uuid: Option<Uuid>,
    pub timestamp: Option<DateTime<FixedOffset>>,
    pub time: Option<Duration>,
    pub tests: usize,
    pub failures: usize,
    pub errors: usize,
    pub test_suites: Vec<TestSuite>,
}

The root element of a JUnit report.

Fields
name: String

The name of this report.

uuid: Option<Uuid>

A unique identifier associated with this report.

This is an extension to the spec that’s used by nextest.

timestamp: Option<DateTime<FixedOffset>>

The time at which the first test in this report began execution.

This is not part of the JUnit spec, but may be useful for some tools.

time: Option<Duration>

The overall time taken by the test suite.

This is serialized as the number of seconds.

tests: usize

The total number of tests from all TestSuites.

failures: usize

The total number of failures from all TestSuites.

errors: usize

The total number of errors from all TestSuites.

test_suites: Vec<TestSuite>

The test suites contained in this report.

Implementations

impl Report

pub fn new(name: impl Into<String>) -> Self

Creates a new Report with the given name.

pub fn set_uuid(&mut self, uuid: Uuid) -> &mut Self

Sets a unique ID for this Report.

This is an extension that’s used by nextest.

pub fn set_timestamp( &mut self, timestamp: impl Into<DateTime<FixedOffset>> ) -> &mut Self

Sets the start timestamp for the report.

pub fn set_time(&mut self, time: Duration) -> &mut Self

Sets the time taken for overall execution.

pub fn add_test_suite(&mut self, test_suite: TestSuite) -> &mut Self

Adds a new TestSuite and updates the tests, failures and errors counts.

When generating a new report, use of this method is recommended over adding to self.TestSuites directly.

pub fn add_test_suites( &mut self, test_suites: impl IntoIterator<Item = TestSuite> ) -> &mut Self

Adds several TestSuites and updates the tests, failures and errors counts.

When generating a new report, use of this method is recommended over adding to self.TestSuites directly.

pub fn serialize(&self, writer: impl Write) -> Result<(), SerializeError>

Serialize this report to the given writer.

pub fn to_string(&self) -> Result<String, SerializeError>

Serialize this report to a string.