Struct quick_junit::TestRerun

pub struct TestRerun {
    pub kind: NonSuccessKind,
    pub timestamp: Option<DateTime<FixedOffset>>,
    pub time: Option<Duration>,
    pub message: Option<String>,
    pub ty: Option<String>,
    pub stack_trace: Option<String>,
    pub system_out: Option<Output>,
    pub system_err: Option<Output>,
    pub description: Option<String>,
}

A rerun of a test.

This is serialized as flakyFailure or flakyError for successes, and as rerunFailure or rerunError for failures/errors.

Fields
kind: NonSuccessKind

The failure kind: error or failure.

timestamp: Option<DateTime<FixedOffset>>

The time at which this rerun began execution.

This is not part of the JUnit spec, but may be useful for some tools.

time: Option<Duration>

The time it took to execute this rerun.

This is not part of the JUnit spec, but may be useful for some tools.

message: Option<String>

The failure message.

ty: Option<String>

The "type" of failure that occurred.

stack_trace: Option<String>

The stack trace, if any.

system_out: Option<Output>

Data written to standard output while the test rerun was executed.

system_err: Option<Output>

Data written to standard error while the test rerun was executed.

description: Option<String>

The description of the failure.

This is serialized and deserialized from the text node of the element.

Implementations

impl TestRerun

pub fn new(kind: NonSuccessKind) -> Self

Creates a new TestRerun of the given kind.

pub fn set_timestamp(
&mut self,
timestamp: impl Into<DateTime<FixedOffset>>
) -> &mut Self

Sets the start timestamp for this rerun.

pub fn set_time(&mut self, time: Duration) -> &mut Self

Sets the time taken for this rerun.

pub fn set_message(&mut self, message: impl Into<String>) -> &mut Self

Sets the message.

pub fn set_type(&mut self, ty: impl Into<String>) -> &mut Self

Sets the type.

pub fn set_stack_trace(&mut self, stack_trace: impl Into<String>) -> &mut Self

Sets the stack trace.

pub fn set_system_out(&mut self, system_out: impl Into<String>) -> &mut Self

Sets standard output.

pub fn set_system_out_lossy(
&mut self,
system_out: impl Into<Vec<u8>>
) -> &mut Self

Sets standard output from a Vec<u8>.

The output is converted to a string, lossily.

pub fn set_system_err(&mut self, system_err: impl Into<String>) -> &mut Self

Sets standard error.

pub fn set_system_err_lossy(
&mut self,
system_err: impl Into<Vec<u8>>
) -> &mut Self

Sets standard error from a Vec<u8>.

The output is converted to a string, lossily.

pub fn set_description(&mut self, description: impl Into<String>) -> &mut Self

Sets the description of the failure.