Crate quick_junit

Generate and read JUnit reports in Rust.

Structs

Output

Represents text that is written out to standard output or standard error during text execution.

Property

Custom properties set during test execution, e.g. environment variables.

Report

The root element of a JUnit report.

TestRerun

A rerun of a test.

Testcase

Represents a single testcase.

Testsuite

Represents a single testsuite.

Enums

Error

The error type used by this crate.

NonSuccessKind

TestcaseStatus

Represents the success or failure of a testcase.

Type Definitions

Result

A specialized Result type where the error is hard-wired to Error.